rollback constants location for circular dependency issue
